+++ relay-log/src/setup.rs
@@ -336,8 +336,7 @@ pub unsafe fn init(config: &LogConfig, sentry: &SentryConfig) {
                 tracing::Level::ERROR | tracing::Level::WARN => {
                     EventFilter::Event | EventFilter::Log
                 }
-                tracing::Level::INFO => EventFilter::Log,
-                tracing::Level::DEBUG | tracing::Level::TRACE => EventFilter::Ignore,
+                _ => EventFilter::Log,
             }),
         )
         .with(match env::var(EnvFilter::DEFAULT_ENV) {