+++ relay-server/src/processing/replays/mod.rs
@@ -21,6 +21,8 @@ pub type Result<T, E = Error> = std::result::Result<T, E>;
 mod filter;
 mod forward;
 mod process;
+#[cfg(feature = "processing")]
+mod store;
 mod validate;
 
 #[derive(Debug, thiserror::Error)]
@@ -31,10 +33,14 @@ pub enum Error {
 
     /// There is an invalid amount of `replay_event` and `replay_recording` items in the envelope.
     ///
-    /// Valid quantity combinations: (0, 0), (1, 1), (1, 0), or (0, 1).
+    /// Valid quantity combinations: (0, 0), (1, 1) or (0, 1).
     /// Standalone events and recordings are supported for SDK compatibility.
     #[error("invalid item count")]
     InvalidItemCount,
+
+    /// Relay event without a recording.
+    #[error("replay event without recording")]
+    EventWithoutRecording,
 
     /// The Replay event could not be parsed from JSON.
     #[error("invalid json: {0}")]
@@ -75,6 +81,16 @@ pub enum Error {
     /// Failed to re-serialize the replay.
     #[error("failed to serialize replay")]
     FailedToSerializeReplay,
+
+    /// Replay recording too large for the consumer.
+    #[cfg(feature = "processing")]
+    #[error("replay recording too large")]
+    TooLarge,
+
+    /// The envelope did not contain an event ID.
+    #[cfg(feature = "processing")]
+    #[error("missing replay ID")]
+    NoEventId,
 }
 
 impl OutcomeError for Error {
@@ -84,6 +100,9 @@ impl OutcomeError for Error {
         let outcome = match &self {
             Self::FilterFeatureFlag => None,
             Self::InvalidItemCount => Some(Outcome::Invalid(DiscardReason::DuplicateItem)),
+            Self::EventWithoutRecording => Some(Outcome::Invalid(
+                DiscardReason::InvalidReplayMissingRecording,
+            )),
             Self::CouldNotParseEvent(_) => {
                 Some(Outcome::Invalid(DiscardReason::InvalidReplayEvent))
             }
@@ -106,6 +125,12 @@ impl OutcomeError for Error {
             }
             Self::Filtered(key) => Some(Outcome::Filtered(key.clone())),
             Self::FailedToSerializeReplay => Some(Outcome::Invalid(DiscardReason::Internal)),
+            #[cfg(feature = "processing")]
+            Self::TooLarge => Some(Outcome::Invalid(DiscardReason::TooLarge(
+                crate::services::outcome::DiscardItemType::ReplayRecording,
+            ))),
+            #[cfg(feature = "processing")]
+            Self::NoEventId => Some(Outcome::Invalid(DiscardReason::Internal)),
         };
         (outcome, self)
     }
@@ -266,11 +291,6 @@ enum ExpandedReplay {
         recording: Bytes,
         video: Bytes,
     },
-    /// A standalone replay event.
-    ///
-    /// Although a `replay_event` and `replay_recording` should always be send together in an
-    /// envelope, this is not the case for some SDKs. As such, support this to not break them.
-    StandaloneEvent { event: Annotated<Replay> },
     /// A standalone replay recording.
     ///
     /// Although a `replay_event` and `replay_recording` should always be send together in an
@@ -284,9 +304,7 @@ impl Counted for ExpandedReplay {
             ExpandedReplay::WebReplay { .. } => {
                 smallvec::smallvec![(DataCategory::Replay, 2)]
             }
-            ExpandedReplay::NativeReplay { .. }
-            | ExpandedReplay::StandaloneEvent { .. }
-            | ExpandedReplay::StandaloneRecording { .. } => {
+            ExpandedReplay::NativeReplay { .. } | ExpandedReplay::StandaloneRecording { .. } => {
                 smallvec::smallvec![(DataCategory::Replay, 1)]
             }
         }
@@ -298,17 +316,15 @@ impl ExpandedReplay {
         match self {
             ExpandedReplay::WebReplay { event, .. } => Some(event),
             ExpandedReplay::NativeReplay { event, .. } => Some(event),
-            ExpandedReplay::StandaloneEvent { event } => Some(event),
             ExpandedReplay::StandaloneRecording { .. } => None,
         }
     }
 
-    fn recording_mut(&mut self) -> Option<&mut Bytes> {
+    fn recording_mut(&mut self) -> &mut Bytes {
         match self {
-            ExpandedReplay::WebReplay { recording, .. } => Some(recording),
-            ExpandedReplay::NativeReplay { recording, .. } => Some(recording),
-            ExpandedReplay::StandaloneEvent { .. } => None,
-            ExpandedReplay::StandaloneRecording { recording } => Some(recording),
+            ExpandedReplay::WebReplay { recording, .. } => recording,
+            ExpandedReplay::NativeReplay { recording, .. } => recording,
+            ExpandedReplay::StandaloneRecording { recording } => recording,
         }
     }
 }